split ui into multiple files
